Figured out how to avoid converting id from long to int

+++ src/main/java/com/PredictionMarket/MarketService/MarketRepository.java
@@ -1,5 +1,5 @@
 package com.PredictionMarket.MarketService;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-public interface MarketRepository extends JpaRepository<Market, Integer> {
+public interface MarketRepository extends JpaRepository<Market, Long> {
 }
